extend H1Client with some customizations
It would be nice to control some behaviour related to the connection
made by the H1Client, as well as the request it sends to establish the
tunnel over that connection.

This adds two optional extensions to the H1Client struct to facilitate
that control:
 - Dialer: a 'standard' net Dialer function which takes the request
   context, the network and the address to dial, returning the resulting
   connection.
 - OnRequest: a function taking the request used to etablish the tunnel,
   invoked just before the request is serialized onto the connection. It
   returns the new request to use (allowing cloning if so desired).

My initial use-cases here for the dialer are to allow optionally dialing through a
CONNECT forward proxy, as well as possibily selecting one of many
endpoints based on some heuristics (e.g. something like happy eyeballs).

OnRequest allows me to add an authorization header primarily, but there
are a few gotchas in my deployment I was able to work around with it.