abstract the conn used by ProxyConnectedSocket
I need to be able to do some things with the connection object used by
the masque proxy. An interface which is the union of net.Conn
and net.PacketConn captures the semantics of net.UDPConn for the most
part. If ProxyConnectedSocket takes that abstraction, I can substitute a
wrapper (or entirely custom) implementation.

The proxy doesn't appear to use any functionality of
net.UDPConn, except for its semantics as a net.Conn, so the abstraction
should be safe and backwards compatible, though there is a risk it
prevents future uses of the underlying net.UDPConn's udp-specific
functions (such as WriteMsgUDP).